requirements.txt
Flask>=3.0.0
SQLAlchemy>=2.0.0
Flask-SQLAlchemy>=3.1.0
python-dotenv>=1.0.0
Flask-Migrate>=4.0.0
Flask-RESTful>=0.3.10
PyJWT>=2.8.0
werkzeug>=3.0.0
gunicorn>=21.2.0
PyPDF2>=3.0.1
pytesseract>=0.3.10
Pillow>=10.0.0
python-docx>=1.1.2
